requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
cymem==2.0.2
cytoolz==0.9.0.1
datefinder==0.7.0
dill==0.2.9
dj-database-url==0.5.0
Django==2.2.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-heroku==0.3.1
django-silk==3.0.2
djangorestframework==3.9.4
# en-core-web-sm==2.0.0
et-xmlfile==1.0.1
gprof2dot==2017.9.19
gunicorn==19.9.0
idna==2.8
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nltk==3.4.1
numpy==1.16.3
openpyxl==2.6.2
pandas==0.24.2
plac==0.9.6
preshed==2.0.1
psycopg2==2.8.2
pycodestyle==2.5.0
Pygments==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
regex==2018.1.10
requests==2.21.0
six==1.12.0
spacy==2.0.18
sqlparse==0.3.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
ujson==1.35
urllib3==1.24.3
whitenoise==4.1.2
wrapt==1.10.11
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz